Practice_2024/requirements.txt
python-telegram-bot==20.0
requests==2.25.1
psycopg2-binary==2.9.3
